[dense] Re-enable SIMD to work with fixed-point
* src/dense/ftdense.c: Use integer SIMD functions for accumulation

* src/dense/ftdense.h: Change types of FT26D6, FT20D12 to better fit
their usage

+++ src/dense/ftdense.h
@@ -20,8 +20,8 @@ extern "C"
 #endif
 
 
-  typedef signed long long FT26D6;            /* 26.6 fixed-point representation  */
-  typedef signed long long FT20D12;           /* 20.12 fixed-point representation  */
+  typedef signed long FT26D6;            /* 26.6 fixed-point representation  */
+  typedef signed int FT20D12;            /* 20.12 fixed-point representation  */
 
   typedef struct
   {
